chroe: remove forest-fil-types crate and relocate items (#2698)
