+++ Makefile
@@ -59,7 +59,6 @@ clean:
 	@cargo clean -p forest_interpreter
 	@cargo clean -p forest_ipld
 	@cargo clean -p forest_json
-	@cargo clean -p forest_fil_types
 	@cargo clean -p forest_rpc
 	@cargo clean -p forest_key_management
 	@cargo clean -p forest_utils
