+++ blockchain/consensus/fil_cns/Cargo.toml
@@ -12,16 +12,17 @@ async-trait.workspace = true
 cid.workspace = true
 fil_actor_interface.workspace = true
 fil_actors_runtime_v9.workspace = true
+filecoin-proofs-api = { workspace = true }
 forest_beacon.workspace = true
 forest_blocks.workspace = true
 forest_chain.workspace = true
 forest_chain_sync.workspace = true
-forest_fil_types.workspace = true
 forest_interpreter.workspace = true
 forest_key_management.workspace = true
 forest_networks.workspace = true
 forest_shim.workspace = true
 forest_state_manager.workspace = true
+forest_utils.workspace = true
 futures.workspace = true
 fvm_ipld_blockstore.workspace = true
 fvm_ipld_encoding.workspace = true
