+++ blockchain/state_manager/Cargo.toml
@@ -14,11 +14,11 @@ byteorder.workspace = true
 cid.workspace = true
 fil_actor_interface.workspace = true
 fil_actors_runtime_v9.workspace = true
+filecoin-proofs-api = { workspace = true }
 forest_beacon.workspace = true
 forest_blocks.workspace = true
 forest_chain.workspace = true
 forest_db.workspace = true
-forest_fil_types.workspace = true
 forest_interpreter.workspace = true
 forest_json.workspace = true
 forest_message = { workspace = true, features = ["blst"] }
