+++ utils/forest_utils/Cargo.toml
@@ -16,7 +16,9 @@ chrono.workspace = true
 cid.workspace = true
 const_format = "0.2.26"
 digest.workspace = true
+filecoin-proofs-api.workspace = true
 flume.workspace = true
+forest_shim.workspace = true
 futures.workspace = true
 fvm_ipld_blockstore.workspace = true
 fvm_ipld_encoding.workspace = true
